Create a build tree function that receives an array
start and end arguments such are needed for recursion.
use recursion until it reaches start greater than end then return null
tree gets the array mid item using the mid equal to the start and end by 2
and use the tree.setLeft and right method to assign left and right of
the current tree in the stack. finally returns the tree.root