Use iteration method together with levelOrder method
within infinite loop
increment height when the node count is not zero thus returns if it's
zero
shift all the queued items and check for left and right nodes to be put
to queue
lessen node count
repeat

in other words, it increments height before even reaching the ends of
the node. help me if i didn't understand this code quite well.